bff-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
